logging: Make output bare of level/target/fields
