+++ src/main.rs
@@ -5,7 +5,7 @@ use nixci::cli;
 #[tokio::main]
 async fn main() -> Result<()> {
     let args = cli::CliArgs::parse();
-    nixci::setup_logging(args.verbose);
+    nixci::logging::setup_logging(args.verbose);
     let _outs = nixci::nixci(args).await?;
     Ok(())
 }
